+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # master
 
+## 1.1.1 / 2011-12-31
+
 * Fix fields being defined for Mongoid / MongoMapper state attributes even if they're already defined in the model
 * Raise error when states / events are referenced in a definition with different types (e.g. both Strings and Symbols)
 * Allow all states / events to be looked up by their string / symbol equivalent